refactor(decks): extract action management logic into reusable ActionListViewModel
