+++ Desktop/Features/Decks/Views/Components/PropertyEditor/PropertyEditorView.xaml
@@ -124,7 +124,7 @@
                 <layout:DeckSidebarHeader Title="Actions" Grid.Row="0">
                     <layout:DeckSidebarHeader.RightContent>
                         <controls:IconButton
-                            Command="{Binding Editor.ClearActionsCommand}"
+                            Command="{Binding Editor.ActionList.ClearActionsCommand}"
                             IconType="ClearAllInk"
                             TooltipText="Clear Actions" />
                     </layout:DeckSidebarHeader.RightContent>
@@ -133,7 +133,7 @@
                 <ae:ActionListView
                     Grid.Row="1"
                     Margin="8"
-                    DataContext="{Binding}" />
+                    DataContext="{Binding Editor.ActionList}" />
             </Grid>
         </Grid>
     </Grid>
